pyproject.toml
[project]
name = "treehacks26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10.32",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "svg.path>=7.0",
]
